fix(mcp): keep stdout protocol-pure, cap stdin during read, share project detection
Three MCP-surface hardening fixes from the full-code audit:

- tracing logs now go to stderr: the default fmt writer is stdout, which
  `icm serve` shares with line-framed JSON-RPC — a WARN emitted in normal
  operation (embedding hiccup, auto-decay failure) could corrupt the MCP
  stream. Verified e2e: stdout stays 100% valid JSON, logs land on stderr.
- the 10 MB stdin line cap is enforced *while reading* (bounded take +
  read_until, remainder drained in bounded chunks) instead of after
  `lines()` had already buffered the whole line — same class of bug as the
  CLI hook-stdin fix (e551c27). Verified e2e: a 12 MB line gets -32600 and
  the server keeps serving.
- project detection is now shared: `project_from_path`/`repo_name_from_url`
  moved to icm-core (`icm_core::project`), used by both the CLI hooks and
  MCP recall. Previously the MCP used the raw cwd basename while hooks
  stored under the git-remote name, so a renamed checkout silently missed
  its recalls. Unit tests (incl. worktree + SSH-remote cases) moved with
  the code.
